+++ app/components/calls_to_action/story_component.html.erb
