+++ app/components/calls_to_action/story_component.rb
@@ -1,0 +1,17 @@
+module CallsToAction
+  class StoryComponent < ViewComponent::Base
+    attr_reader :name, :heading, :image, :link, :text
+
+    def initialize(name:, heading:, image:, link:, text:)
+      @name    = name
+      @heading = heading
+      @image   = image
+      @link    = link
+      @text    = text
+    end
+
+    def button
+      link_to(tag.span("Read #{name}'s story"), link, class: "call-to-action__contents__button")
+    end
+  end
+end
